requirements.txt
django
djangorestframework
psycopg2-binary
pillow
python-decouple
python-slugify
drf-yasg

